Save more informations in some forge objects
In order to improve user experience and support new features like
code-review or replies, some informations need to be added in
issue / pullrequest objects and a new object pullreq-version has been
created.

* issue-post object:
  -> thread-id: id of a thread discussion
  -> reply-to:  id of the initial post in reply to

* pullreq object:
  -> versions: list of version object

* pullreq-post object:
  -> thread-id:   id of a thread discussion
  -> diff-p:      describe if this post is a diff post
  -> resolved-by: the user who resolved this thread
  -> reply-to:    id of the initial post in reply to
  -> head-ref:    head SHA for this post
  -> commit-ref:  commit SHA for this post
  -> base-ref:    base SHA for this post
  -> path:        file path where the diff is posted
  -> old-line:    the line where the diff is posted before changes
  -> new-line:    the line where the diff is posted after changes

* pullreq-version: contains informations like the base and head SHA
reference of a pullreq version.
These informations are needed for eg. when we want to know for which
version of a pullreq are diff posts.

If an old forge database is detected, it will be upgraded to the new
version 8.

Signed-off-by: Julien Masson <massonju.eseo@gmail.com>