Clean-up "comment" functions
In the code all the "comment" are treated as a post object.
A post object can be found in a pull-request or an issue for eg.

To avoid confusion it makes more sense to use "post" instead of
"comment" in the function name when handling post object.

Also forge-comment-at-point function has been removed since it's not
used.

Signed-off-by: Julien Masson <massonju.eseo@gmail.com>